refactor: remove index file from events module and update event imports
