--- backend/src/events/index.ts
@@ -1,4 +1,0 @@
-// Re-export all event types
-export * from './message.events';
-export * from './contact.events';
-export * from './user.events';
